Implemented Standard Hyena Enemy
Implemented the AI for the standard Hyena enemy including patrolling, searching for the player if it lost them, chasing the player and bait, eating bait, ignoring the player when there is bait in front of it, and attacking the player. There are no visual indications that the enemy has shifted between searching, patrolling, or chasing after the player. Also the enemy will not ignore the player if they're hiding because hiding hasn't been implemented yet.