Housekeeping Update
Fixed the problem with the audio manager, created a prefab for the UI for the game play screens, add numerous art assets in to the game, created prefabs for all the environment art to make it easier to create levels, implemented basic sprites for the player, enemies and obstacles, remove unnecessary scripts and 'better' organized the assets folder.
Note: due to changes the old levels will not be playable and the noodle birds will probably have bugs to to their attacks being reworked. I will remove the old levels once I implement the new levels and I will fix the noodle birds when I get around to implementing their animations.