Obstacle Update
Updated the rock, screamer plant, and quicksand obstacle prefabs to work as intended. Altered the noodle birds colliders to give teh player more leeway. Lastly, Updated the gameplay UI to have bait icons (Also made it so that the bait Icon is set to the specific berry for that level).