State Indicators Update
Updated the art for the state indicators to reflect the more angular art style we were going for as well as made it so that it goes from yellow to orange to red based on the dangerousness of the enemy. Created overlays for all three parts of the level to better tell apart the background and foreground. Tweaked the camera follow to more smoothly follow the player.