Fixed SFX Range Issue
Fixed the sound effects so that you no longer hear enemies that are miles away from you and the audio will properly get quieter the farther you are away from the audio source. I created a public float variable to manage the max distance and made it so that it can never be smaller than 1 and no larger than 25. I set the default to 15 unity units. Note: the sound effects are noticeably quieter now and I plan to fix that issue once the new normalized sounds get put into the game.