Game State Flow Update
Updated the flow of the game state by implementing quit confirmation and a sound button to the pause menu (help button still does not work), added death animation and made the game wait for it to play out before popping up the game over screen, and modified the level complete screen to show the eggs the player collected. Added a overlay to appear when the pause menu or popup menu appears in the game.